+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Enhancements
+
+* [Braintree.HTTP] Support both access_token and public/private keys usage in configuration
+
 ## v0.9.0 2018-06-18
 
 ### Enhancements